Move `ctrl` to `om` component
