+++ components/bbc/src/brake.c
@@ -54,12 +54,12 @@ static void brake_100Hz()
 
 	bool brake_authorized = CANRX_is_message_SUP_Authorization_ok()
 	    && CANRX_get_SUP_brakeAuthorized()
-	    && CANRX_is_message_CTRL_VelocityCommand_ok();
+	    && CANRX_is_message_OM_VelocityCommand_ok();
 
 	float cmd;
 
 	if (brake_authorized) {
-		cmd = ((float32_t) CANRX_get_CTRL_brakePercent()) / 100.0;
+		cmd = ((float32_t) CANRX_get_OM_brakePercent()) / 100.0;
 		base_request_state(SYS_STATE_DBW_ACTIVE);
 	} else {
 		cmd = 0.0;
