+++ components/om/component.toml
@@ -1,0 +1,3 @@
+[metadata]
+name = 'om'
+description = 'velocity controller'
