+++ components/om/src/firmware-base
@@ -1,0 +1,1 @@
+../../../lib/firmware-base
