+++ components/om/src/om.h
@@ -1,0 +1,4 @@
+#ifndef OM_H
+#define OM_H
+
+#endif	// OM_H
