--- dbw/node_fw/mod/ctrl/ctrl.h
@@ -1,4 +1,0 @@
-#ifndef CTRL_H
-#define CTRL_H
-
-#endif
